made components for numbers, guesses
